+++ src/Glpi/Api/HL/Router.php
@@ -95,7 +95,7 @@ use function Safe\preg_match;
 class Router
 {
     /** @var string */
-    public const API_VERSION = '2.3.0';
+    public const API_VERSION = '2.4.0';
 
     /**
      * @var AbstractController[]
@@ -187,8 +187,8 @@ EOT;
             ],
             [
                 'api_version' => '2',
-                'version' => '2.3.0',
-                'endpoint' => $CFG_GLPI['url_base'] . '/api.php/v2.3',
+                'version' => '2.4.0',
+                'endpoint' => $CFG_GLPI['url_base'] . '/api.php/v2.4',
             ],
         ];
     }